Grammar Update
The code can plot the L2 norm correctly. Rearranged inputs of functions being called and changed how the support size was found. Moved GaussPointsandWeights to the top, and had to make the node interval an input from information in ConvergenceStudies so that the Gauss evaluation points can change with the change in nodal interval size.